@@ -6,7 +6,7 @@ readme = "README.md"
 authors = [
     { name = "Ben-Cassese", email = "b.c.cassese@columbia.edu" }
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 classifiers = [
     "Intended Audience :: Science/Research",
     "Intended Audience :: Developers",
@@ -14,7 +14,6 @@ classifiers = [
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -26,7 +25,6 @@ dependencies = [
     "astropy>=6.0.1",
     "astroquery>=0.4.7",
     "chex>=0.1.88",
-    "emcee>=3.1.6",
     "ipykernel>=6.29.5",
     "jax>=0.4.30",
     "jplephem>=2.22",
@@ -65,8 +63,13 @@ dev = [
     "lightkurve>=2.5.0",
     "spiceypy>=6.0.0",
     "tess-prf>=0.1.3",
-    "rebound>=4.4.10",
-    "reboundx>=4.4.1",
+    "setuptools>=75.0.0",
+    "rebound>=4.5.1",
+    "reboundx>=4.6.0",
+    "lxml>=6.0.2",
+    "emcee>=3.1.6",
+    "corner>=2.2.3",
+    "assist>=1.1.11",
 ]
 
 [project.urls]
@@ -88,6 +91,9 @@ exclude = [
     "scratch*",
 ]
 pydocstyle.convention = "google"
+
+[tool.uv]
+no-build-isolation-package = ["rebound", "reboundx"]
 
 [tool.codespell]
 skip = "./docs/_build/*, **/*.ipynb, ./src/jorbit/data/observatory_codes.py"
