@@ -15,6 +15,11 @@ def newtonian_gravity(inputs: SystemState) -> jnp.ndarray:
     Set up to separate massive and tracer particles, so systems with many tracers
     will not compute useless pairwise interactions.
 
+    Note: We use "stop_gradient" on perturbers that are passed as fixed inputs, so
+    any gradients with respect to these perturber quantities will not be correct. To
+    track gradients with respect to perturbers, they must be included as "massive"
+    particles, not "fixed perturbers".
+
     Args:
         inputs (SystemState): The instantaneous state of the system.
 
@@ -24,42 +29,53 @@ def newtonian_gravity(inputs: SystemState) -> jnp.ndarray:
             first followed by tracer particles.
     """
     M = inputs.massive_positions.shape[0]  # number of massive particles
+    gms = jnp.exp(inputs.log_gms)  # (M,)
 
-    # 1. Compute accelerations on massive particles due to other massive particles
+    # massive particles due to other massive particles
     dx_massive = (
         inputs.massive_positions[:, None, :] - inputs.massive_positions[None, :, :]
     )  # (M,M,3)
     r2_massive = jnp.sum(dx_massive * dx_massive, axis=-1)  # (M,M)
     r3_massive = r2_massive * jnp.sqrt(r2_massive)  # (M,M)
 
-    # Mask for i!=j calculations among massive particles
     mask_massive = ~jnp.eye(M, dtype=bool)  # (M,M)
     prefac_massive = jnp.where(mask_massive, 1.0 / r3_massive, 0.0)
 
-    # Accelerations on massive particles from massive particles
     a_massive = -jnp.sum(
-        prefac_massive[:, :, None]
-        * dx_massive
-        * jnp.exp(inputs.log_gms[None, :, None]),
+        prefac_massive[:, :, None] * dx_massive * gms[None, :, None],
         axis=1,
     )  # (M,3)
 
-    # 2. Compute accelerations on tracer particles due to massive particles
-    # (This will work even when T=0 due to JAX's shape polymorphism)
+    # tracer particles due to massive particles
     dx_tracers = (
         inputs.tracer_positions[:, None, :] - inputs.massive_positions[None, :, :]
     )  # (T,M,3)
     r2_tracers = jnp.sum(dx_tracers * dx_tracers, axis=-1)  # (T,M)
     r3_tracers = r2_tracers * jnp.sqrt(r2_tracers)  # (T,M)
 
-    # Accelerations on tracer particles from massive particles
     a_tracers = -jnp.sum(
-        (1.0 / r3_tracers)[:, :, None]
-        * dx_tracers
-        * jnp.exp(inputs.log_gms[None, :, None]),
+        (1.0 / r3_tracers)[:, :, None] * dx_tracers * gms[None, :, None],
         axis=1,
     )  # (T,3)
 
-    # Combine accelerations for all particles
-    # This works even when T=0 thanks to JAX's handling of zero-sized arrays
-    return jnp.concatenate([a_massive, a_tracers], axis=0)
+    all_as = jnp.concatenate([a_massive, a_tracers], axis=0)
+
+    # Fixed perturbers acting on all (massive + tracer) particles.
+    # stop_gradient since we never need gradients through fixed perturber quantities.
+    p_pos = jax.lax.stop_gradient(inputs.fixed_perturber_positions)
+    p_gms = jax.lax.stop_gradient(jnp.exp(inputs.fixed_perturber_log_gms))
+
+    dx_perturbers = (
+        jnp.concatenate([inputs.massive_positions, inputs.tracer_positions], axis=0)[
+            :, None, :
+        ]
+        - p_pos[None, :, :]
+    )  # (N,P,3)
+    r2_perturbers = jnp.sum(dx_perturbers * dx_perturbers, axis=-1)  # (N,P)
+    r3_perturbers = r2_perturbers * jnp.sqrt(r2_perturbers)  # (N,P)
+    a_perturbers = -jnp.sum(
+        (1.0 / r3_perturbers)[:, :, None] * dx_perturbers * p_gms[None, :, None],
+        axis=1,
+    )  # (N,3)
+
+    return all_as + a_perturbers
