@@ -182,7 +182,10 @@ def ias15_step(
 ) -> SystemState:
     """Take a single step using the IAS15 integrator.
 
-    Contains all of the predictor/corrector logic and step validity checks.
+    Contains all of the predictor/corrector logic and step validity checks. Does not
+    accept any pre-computed perturber information, since we don't know the times this
+    will be needed until runtime. For a static version that accepts pre-computed
+    perturber data, see ias15_static_step.
 
     Args:
         initial_system_state (SystemState):
@@ -236,7 +239,9 @@ def ias15_step(
     ) -> tuple:
         predictor_corrector_error_last = predictor_corrector_error
         predictor_corrector_error = 0.0
-        for n, h, c, r in zip(range(1, 8), IAS15_H[1:], IAS15_sub_cs, IAS15_sub_rs):
+        for n, h, c, r in zip(
+            range(1, 8), IAS15_H[1:], IAS15_sub_cs, IAS15_sub_rs, strict=True
+        ):
             step_time = t_beginning + dt * h
             x, v = _estimate_x_v_from_b(
                 a0=a0,
@@ -246,6 +251,8 @@ def ias15_step(
                 dt=dt,
                 bp=b[::-1],
             )
+            # note that the fixed perturber bits likely can/will be overwritten by the
+            # acceleration function- see ias15_static_step + create_static_default_acceleration_func
             acc_state = SystemState(
                 massive_positions=x[:M],
                 massive_velocities=v[:M],
@@ -253,6 +260,13 @@ def ias15_step(
                 tracer_velocities=v[M:],
                 log_gms=initial_system_state.log_gms,
                 time=step_time,
+                fixed_perturber_positions=jnp.empty(
+                    (0, 3),
+                ),
+                fixed_perturber_velocities=jnp.empty(
+                    (0, 3),
+                ),
+                fixed_perturber_log_gms=jnp.empty((0,)),
                 acceleration_func_kwargs=initial_system_state.acceleration_func_kwargs,
             )
             at = acceleration_func(acc_state)
@@ -362,6 +376,9 @@ def ias15_step(
         tracer_velocities=v0[M:],
         log_gms=initial_system_state.log_gms,
         time=t_beginning + dt_done,
+        fixed_perturber_positions=initial_system_state.fixed_perturber_positions * 0,
+        fixed_perturber_velocities=initial_system_state.fixed_perturber_velocities * 0,
+        fixed_perturber_log_gms=initial_system_state.fixed_perturber_log_gms * 0,
         acceleration_func_kwargs=initial_system_state.acceleration_func_kwargs,
     )
 
@@ -478,7 +495,7 @@ def ias15_evolve(
             )
         )
         # jax.debug.print(
-        #     "finished taking steps to goal time in {x} iterations", x=iter_num
+        #     "finished taking steps to goal time in {x} iterations", x=_iter_num
         # )
 
         return (final_system_state, final_integrator_state)
